Modify serializers, update requirements
Add BaseSerializer as parent for others serializers
Little code change in serializers
remove ModelSerializer.py
Update requirements.txt
