+++ assets/serializers/ModelsSerializers.py
@@ -1,23 +1,21 @@
 from ..databases.fields.DatabaseFields import *
 from ..databases.fields.FieldsValues import NOT_PROVIDED
 from assets.databases.models.ShopsModel import ShopsModel
+from .BaseSerializer import BaseSerializer
 
 
-class ModelSerializer:
+class ModelSerializer(BaseSerializer):
     def __init__(self, model, data: dict):
-        self.is_valid = True
-        self.errors = ''
+        super().__init__()
         self._model = model
         self._input_data = data
         self._table_fields = [x for x in dir(model) if not callable(getattr(model, x)) and not x.startswith("__")
                               and not x.startswith("_")]
         self.data = None
-        self.serialize()
+        self.__serialize_data()
 
-    def serialize(self):
-        self.__check_not_null_fields()
-
-        if self.errors:
+    def __serialize_data(self):
+        if not self.__are_not_null_fields_correct():
             return False
 
         for key, value in self._input_data.items():
@@ -34,7 +32,7 @@ class ModelSerializer:
         self.data = self._input_data
         return True
 
-    def __check_not_null_fields(self):
+    def __are_not_null_fields_correct(self):
         not_null_fields = [x for x in self._table_fields if getattr(self._model, x).not_null
                            and getattr(self._model, x).primary is False]
         missing_fields = [x for x in self._table_fields if x not in self._input_data.keys()
@@ -46,15 +44,12 @@ class ModelSerializer:
         empty_not_null_fields = [x for x in empty_values if x in not_null_fields]
 
         if empty_not_null_fields:
-            self._set_error(f"This fields reqired values: {empty_not_null_fields}")
+            self._set_error(f"This fields required values: {empty_not_null_fields}")
 
         if missing_fields:
             self._set_error(f"Missing fields {missing_fields}")
 
-    def _set_error(self, message):
-        if self.is_valid is True:
-            self.is_valid = False
-        self.errors = f"{self.errors}; {message}" or message
+        return True if self.is_valid else False
 
 
 class ShopsSerializer(ModelSerializer):
