+++ requirements.txt
@@ -1,7 +1,12 @@
+aniso8601==8.0.0
 beautifulsoup4==4.8.2
+click==7.1.2
 Flask==1.1.1
 Flask-Cors==3.0.8
 Flask-RESTful==0.3.8
+itsdangerous==1.1.0
+Jinja2==2.11.2
+MarkupSafe==1.1.1
 mysql-connector-python==8.0.20
 python-dotenv==0.13.0
 pytz==2019.3
